refactor: Separate Nav.jsx & Madal.jsx
